Restructured work history component folder added
All restructured work history components have now been moved from
the top level component folder into a specific restructured work
history one. This is the same for the spec files. The individual
files have also been renamed to reflect this change.
