+++ app/components/candidate_interface/restructured_work_history/gap_component.html.erb
