+++ app/components/candidate_interface/restructured_work_history/gap_component.rb
@@ -1,0 +1,11 @@
+module CandidateInterface
+  module RestructuredWorkHistory
+    class GapComponent < ViewComponent::Base
+      include ViewHelper
+
+      def initialize(break_period:)
+        @break_period = break_period
+      end
+    end
+  end
+end
