+++ app/components/candidate_interface/restructured_work_history/job_component.html.erb
