--- app/components/candidate_interface/restructured_work_history_gap_component.rb
@@ -1,9 +1,0 @@
-module CandidateInterface
-  class RestructuredWorkHistoryGapComponent < ViewComponent::Base
-    include ViewHelper
-
-    def initialize(break_period:)
-      @break_period = break_period
-    end
-  end
-end
