+++ app/views/candidate_interface/restructured_work_history/review/show.html.erb
@@ -21,7 +21,7 @@
 <% unless @application_form.can_complete? && @application_form.application_work_experiences.blank? %>
   <section class="app-summary-card govuk-!-margin-bottom-6">
     <div class="app-summary-card__body">
-      <%= render(CandidateInterface::RestructuredWorkHistoryReviewComponent.new(application_form: @application_form)) %>
+      <%= render(CandidateInterface::RestructuredWorkHistory::ReviewComponent.new(application_form: @application_form)) %>
     </div>
   </section>
